+++ packages/base/CHANGELOG.md
@@ -15,6 +15,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [5.4.6] - 2020-11-15
+### Added
+ - `DBConnection.selectParsedAsObject` for Postgres warning added: 
+  Postgres return all field name in lower case if `AS "normalName"` is not specified, so better to write a query as such
+  `select ID as "ID", modelName AS "modelName" from ..` instead of `select ID, modelName from ..`
+
+### Changed
+ - `ServerRepository.selectScalar()` optimized to use less JS memory
+
+## [5.4.5] - 2020-11-14
+### Added
+ - `ServerRepository.selectAsArrayOfValues()` - for Repository with ONE attribute returns a flat array of attribute values
+   ```javascript
+    const usersIDs = UB.Repository('uba_user').attrs('ID').limit(100).selectAsArrayOfValues()
+    // usersIDs is array of IDs [1, 2, 3, 4]
+   ```
+   
+## [5.4.4] - 2020-11-12
 ## [5.4.3] - 2020-11-10
 ### Added
  - `argv.establishConnectionFromCmdLineAttributes`: default value for -host parameter is changed to `auto`.